jxl-color: SIMD PQ transfer function
